When a fact is deleted, the indexing changes

+++ Streetcode/Streetcode.BLL/MediatR/Streetcode/Fact/Delete/DeleteFactHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using FluentResults;
 using MediatR;
+using Streetcode.BLL.Dto.Streetcode.TextContent.Fact;
 using Streetcode.BLL.Interfaces.Logging;
+using Streetcode.BLL.Util;
 using Streetcode.DAL.Repositories.Interfaces.Base;
 
 namespace Streetcode.BLL.MediatR.Streetcode.Fact.Delete
@@ -28,7 +30,25 @@ namespace Streetcode.BLL.MediatR.Streetcode.Fact.Delete
                 return Result.Fail(errorMsg);
             }
 
+            // Getting all the facts related to the same Streetcode
+            var facts = await _repositoryWrapper.FactRepository
+                .GetAllAsync(f => f.StreetcodeId == fact.StreetcodeId);
+
+            if (facts == null || !facts.Any())
+            {
+                string errorMsg = $"No facts found for StreetcodeId: {fact.StreetcodeId}";
+                _logger.LogError(request, errorMsg);
+                return Result.Fail(new Error(errorMsg));
+            }
+
             _repositoryWrapper.FactRepository.Delete(fact);
+
+            // Use the function to update the order of facts
+            FactOrderHelper.UpdateFactOrder(facts.ToList(), fact.Id);
+
+            // Saving changes
+            _repositoryWrapper.FactRepository.UpdateRange(facts);
+
             var resultIsSuccess = await _repositoryWrapper.SaveChangesAsync() > 0;
             if (resultIsSuccess)
             {
